restructured folders in api call client side
