+++ client/src/api/services/course/courseService.ts
@@ -16,48 +16,23 @@ export const getCoursesByInstructorService = async (endpoint: string) => {
   return response.data;
 };
 
-export const getLessonsByCourseService = async (
+export const getAllCoursesService = async (endpoint: string) => {
+  const response = await api.get(
+    `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}`
+  );
+  return response;
+};
+
+export const getIndividualCourseService = async (
   endpoint: string,
   courseId: string
 ) => {
   const response = await api.get(
     `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${courseId}`
   );
-  return response.data;
+  return response;
 };
 
-export const addLessonsService = async (
-  endpoint: string,
-  courseId: string,
-  lesson: FormData
-) => {
-  const response = await api.post(
-    `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${courseId}`,
-    lesson
-  );
-  return response.data;
-};
 
-export const getLessonsByIdService = async (
-  endpoint: string,
-  lessonId: string
-) => {
-  const response = await api.get(
-    `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${lessonId}`
-  );
-  return response.data;
-};
 
-export const getCloudFrontVideoUrlService = async (endpoint:string,key:string)=>{
-  const response = await api.get(
-    `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${key}`
-  );
-  return response.data;
-}
 
-export const getQuizzesByLessonService = async(endpoint:string,lessonId:string)=>{
-  const response = await api.get(
-    `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${lessonId}`
-  );
-  return response.data;
-}
