+++ client/src/api/services/course/quizService.ts
@@ -1,0 +1,16 @@
+import CONSTANTS_COMMON from "../../../constants/common";
+import api from "../../middlewares/interceptors";
+
+export const getCloudFrontVideoUrlService = async (endpoint:string,key:string)=>{
+    const response = await api.get(
+      `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${key}`
+    );
+    return response.data;
+  }
+  
+  export const getQuizzesByLessonService = async(endpoint:string,lessonId:string)=>{
+    const response = await api.get(
+      `${CONSTANTS_COMMON.API_BASE_URL}/${endpoint}/${lessonId}`
+    );
+    return response.data;
+  }
