+++ client/src/components/pages/Course/WatchLessons.tsx
@@ -3,10 +3,10 @@ import VideoPlayer from "./VideoPlayer";
 import AboutLesson from "./AboutLesson";
 import Quizzes from "./Quizzes";
 import Discussion from "./Discussion";
-import { getLessonById } from "../../../api/endpoints/course/course";
 import { useParams } from "react-router-dom";
 import { toast } from "react-toastify";
-import { getLessonsByCourse } from "../../../api/endpoints/course/course";
+import { getLessonById } from "../../../api/endpoints/course/lesson";
+import { getLessonsByCourse } from "../../../api/endpoints/course/lesson";
 import { ApiResponseLesson } from "../../../api/types/apiResponses/apResponseLesson";
 import { Media } from "../../../api/types/apiResponses/apResponseLesson";
 import { BiVideo } from "react-icons/bi";
