+++ client/src/components/pages/add-lessons/AddLessonsForm.tsx
@@ -14,7 +14,7 @@ import * as Yup from "yup";
 import QuizSwitch from "./QuizSwitch";
 import { Tooltip } from "@material-tailwind/react";
 import { QuizzesComponent } from "./QuizesComponent";
-import { addLesson } from "../../../api/endpoints/course/course";
+import { addLesson } from "../../../api/endpoints/course/lesson";
 import { FormValuesLesson } from "../../../types/lesson";
 import SpinnerDialog from "../../common/spinner";
